perf(client): load highlight.js lazily
highlight.js (~330 KB) was loaded at startup through two chains:

- desktop.ts -> note_tooltip -> content_renderer -> syntax_highlight, which
  imported the library statically. The highlighting entry points are all
  async, so the library is now resolved via dynamic imports when
  highlighting actually runs. This also fixes formatCodeBlocks not awaiting
  the per-block highlighting it triggers.
- The always-loaded highlights sidebar imported HighlightsListOptions from
  the text note options page, which statically imports highlight.js for its
  theme list. The component now lives in its own module.
